Cumulative updates from pre-production

+++ migrations/versions/98c3d8971d70_fifth_update.py
@@ -1,4 +1,4 @@
-"""Fifth Update.
+"""Fifth Update - Add template_parameters, template_metadata cols to deployments
 
 Revision ID: 98c3d8971d70
 Revises: 98c3d8971d6f
@@ -6,7 +6,7 @@ Create Date: 2020-04-09
 """
 from alembic import op
 import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
+from app.lib import alembic_helper
 
 # revision identifiers, used by Alembic.
 revision = '98c3d8971d70'
@@ -16,12 +16,16 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('deployments', sa.Column('template_parameters', sa.Text, nullable=True))
-    op.add_column('deployments', sa.Column('template_metadata', sa.Text, nullable=True))
+    if not alembic_helper.column_exists('deployments', 'template_parameters'):
+        op.add_column('deployments', sa.Column('template_parameters', sa.Text, nullable=True))
+    if not alembic_helper.column_exists('deployments', 'template_metadata'):
+        op.add_column('deployments', sa.Column('template_metadata', sa.Text, nullable=True))
     # ### end Alembic commands ###
 
 
 def downgrade():
-    op.drop_column('deployments', 'template_parameters')
-    op.drop_column('deployments', 'template_metadata')
+    if alembic_helper.column_exists('deployments', 'template_parameters'):
+        op.drop_column('deployments', 'template_parameters')
+    if alembic_helper.column_exists('deployments', 'template_metadata'):
+        op.drop_column('deployments', 'template_metadata')
     # ### end Alembic commands ###
